+++ xugrid/ugrid/ugrid1d.py
@@ -1,5 +1,5 @@
 from itertools import chain
-from typing import Any, Dict, Sequence, Tuple, Union
+from typing import Any, Dict, Optional, Sequence, Tuple, Union
 
 import numpy as np
 import pandas as pd
@@ -704,7 +704,10 @@ class Ugrid1d(AbstractUgrid):
         )
 
     def refine_by_vertices(
-        self, vertices: FloatArray, return_index: bool = False
+        self,
+        vertices: FloatArray,
+        return_index: bool = False,
+        tolerance: Optional[float] = None,
     ) -> "Ugrid1d":
         """
         Refine Ugrid1d with extra vertices to be inserted and returns new grid.
@@ -719,6 +722,12 @@ class Ugrid1d(AbstractUgrid):
         return_index: bool, optional
             If set to to True, the index of the new vertices in the grid will be
             returned. Defaults to False.
+        tolerance: float, optional
+            The tolerance used to determine whether a point is on an edge. This
+            accounts for the inherent inexactness of floating point calculations.
+            If None, an appropriate tolerance is automatically estimated based on
+            the geometry size. Consider adjusting this value if edge detection
+            results are unsatisfactory.
 
         Returns
         -------
@@ -748,7 +757,7 @@ class Ugrid1d(AbstractUgrid):
         >>> print(new.node_coordinates[new_vertices_index])
 
         """
-        edge_index = self.celltree.locate_points(vertices)
+        edge_index = self.celltree.locate_points(vertices, tolerance)
         invalid = edge_index == -1
         if invalid.any():
             raise ValueError(