+++ xugrid/ugrid/ugrid2d.py
@@ -971,7 +971,9 @@ class Ugrid2d(AbstractUgrid):
         )[0]
 
     def compute_barycentric_weights(
-        self, points: FloatArray
+        self,
+        points: FloatArray,
+        tolerance: Optional[float] = None,
     ) -> Tuple[IntArray, FloatArray]:
         """
         Find in which face the points are located, and compute the barycentric
@@ -980,13 +982,19 @@ class Ugrid2d(AbstractUgrid):
         Parameters
         ----------
         points: ndarray of floats with shape ``(n_point, 2)``
+        tolerance: float, optional
+            The tolerance used to determine whether a point is on an edge. This
+            accounts for the inherent inexactness of floating point calculations.
+            If None, an appropriate tolerance is automatically estimated based on
+            the geometry size. Consider adjusting this value if edge detection
+            results are unsatisfactory.
 
         Returns
         -------
         face_index: ndarray of integers with shape ``(n_points,)``
         weights: ndarray of floats with shape ```(n_points, n_max_node)``
         """
-        return self.celltree.compute_barycentric_weights(points)
+        return self.celltree.compute_barycentric_weights(points, tolerance)
 
     def rasterize_like(
         self, x: FloatArray, y: FloatArray